+++ app/Http/Controllers/Models/CustomerController.php
@@ -2,6 +2,9 @@
 
 namespace App\Http\Controllers\Models;
 
+use App\Events\Customer\CustomerCreatedEvent;
+use App\Events\Customer\CustomerEditedEvent;
+use App\Events\Customer\CustomerRemovedEvent;
 use App\Http\Controllers\Controller;
 use App\Models\Address;
 use App\Models\AddressParent;
@@ -48,6 +51,7 @@ class CustomerController extends Controller
             'passport_first_name' => $request->input('passport_first_name'),
             'passport_middle_name' => $request->input('passport_middle_name'),
             'passport_last_name' => $request->input('passport_last_name'),
+            'passport_country_of_issue' => $request->input('passport_country_of_issue'),
             'passport_number' => $request->input('passport_number'),
             'passport_issue_date' => $request->input('passport_issue_date'),
             'passport_expiry_date' => $request->input('passport_expiry_date'),
@@ -57,7 +61,7 @@ class CustomerController extends Controller
             'loyalty_number' => $request->input('loyalty_number'),
         ]);
         $homeAddress = Address::create([
-            'name' => $request->input('title') . ' ' . $request->input('first_name') . ' ' .  $request->input('last_name'),
+            'name' => $request->input('title') . ' ' . $request->input('first_name') . ' ' . $request->input('last_name'),
             'address_parent_id' => AddressParent::getParentId('customer'),
             'address_line_1' => $request->input('home_address_line_1'),
             'address_line_2' => $request->input('home_address_line_2'),
@@ -69,10 +73,9 @@ class CustomerController extends Controller
         $customer->home_address_id = $homeAddress->id;
         if ($request->input('home_is_billing') == 'on') {
             $customer->billing_address_id = LocationsRepository::cloneAddressToAddress($homeAddress, AddressParent::getParentId('customer'))->id;
-        }
-        else {
+        } else {
             $billingAddress = Address::create([
-                'name' => $request->input('title') . ' ' . $request->input('first_name') . ' ' .  $request->input('last_name'),
+                'name' => $request->input('title') . ' ' . $request->input('first_name') . ' ' . $request->input('last_name'),
                 'address_parent_id' => AddressParent::getParentId('customer'),
                 'address_line_1' => $request->input('billing_address_line_1'),
                 'address_line_2' => $request->input('billing_address_line_2'),
@@ -83,10 +86,11 @@ class CustomerController extends Controller
             ]);
             $customer->billing_address_id = $billingAddress->id;
         }
-        if ($request->has('profile_picture')  && $request->file('profile_picture') != null) {
+        if ($request->has('profile_picture') && $request->file('profile_picture') != null) {
             $customer->profile_picture = $request->file('profile_picture')->storePublicly('uploads/images/customers');
         }
         $customer->save();
+        event(new CustomerCreatedEvent($customer));
         return redirect()->route('customers.view', ['customer' => $customer,]);
     }
 
@@ -123,6 +127,7 @@ class CustomerController extends Controller
             'passport_number' => $request->input('passport_number'),
             'passport_issue_date' => $request->input('passport_issue_date'),
             'passport_expiry_date' => $request->input('passport_expiry_date'),
+            'passport_country_of_issue' => $request->input('passport_country_of_issue'),
             't_shirt_size_id' => $request->input('t_shirt_size_id'),
             'hat_size_id' => $request->input('hat_size_id'),
             'notes' => $request->input('notes'),
@@ -139,8 +144,7 @@ class CustomerController extends Controller
         $customer->homeAddress->save();
         if ($request->input('home_is_billing') == 'on') {
             LocationsRepository::cloneAddressToAddress($customer->homeAddress, AddressParent::getParentId('customer'), $customer->billingAddress);
-        }
-        else {
+        } else {
             $customer->billingAddress->update([
                 'address_line_1' => $request->input('billing_address_line_1'),
                 'address_line_2' => $request->input('billing_address_line_2'),
@@ -151,16 +155,21 @@ class CustomerController extends Controller
             ]);
             $customer->billingAddress->save();
         }
-        if ($request->has('profile_picture')  && $request->file('profile_picture') != null) {
+        if ($request->has('profile_picture') && $request->file('profile_picture') != null) {
             $customer->profile_picture = $request->file('profile_picture')->storePublicly('uploads/images/customers');
         }
         $customer->save();
+        event(new CustomerEditedEvent($customer));
         return redirect()->route('customers.view', ['customer' => $customer,]);
     }
 
     public function destroy(Customer $customer)
     {
+        if ($customer->orderCustomers()->count() > 0) {
+            return back()->withErrors(trans('custom.used-elsewhere', ['model' => 'Customer', 'parent' => 'Order']));
+        }
         $customer->delete();
+        event(new CustomerRemovedEvent($customer));
         return redirect()->route('customers.all');
     }
 }
