+++ app/Http/Controllers/Models/TourController.php
@@ -22,7 +22,7 @@ class TourController extends Controller
 
     public function store(Request $request)
     {
-        $request->validate(Tour::RULES);
+        $request->validate(Tour::getValidationRules());
         $tour = Tour::create([
             'event_id' => $request->input('event_id'),
             'name' => $request->input('name'),
@@ -35,10 +35,11 @@ class TourController extends Controller
             'stock_control_active' => $request->input('stock_control_active') === 'on' ? 1 : 0,
             'stock' => $request->input('stock'),
             'booking_form_url' => $request->input('booking_form_url'),
-            'tour_colour_id' => $request->input('tour_colour_id'),
+            'tour_category_id' => $request->input('tour_category_id'),
             'deposit' => $request->input('deposit'),
             'is_active' => $request->input('is_active') === 'on' ? 1 : 0,
             'notes' => $request->input('notes'),
+            'invoice_footer' => $request->input('deposit'),
         ]);
         return redirect()->route('tours.view', ['tour' => $tour,]);
     }
@@ -55,7 +56,7 @@ class TourController extends Controller
 
     public function update(Request $request, Tour $tour)
     {
-        $request->validate(Tour::RULES);
+        $request->validate(Tour::getValidationRules());
         $tour->update([
             'event_id' => $request->input('event_id'),
             'name' => $request->input('name'),
@@ -69,15 +70,19 @@ class TourController extends Controller
             'stock_control_active' => $request->input('stock_control_active') === 'on' ? 1 : 0,
             'stock' => $request->input('stock'),
             'booking_form_url' => $request->input('booking_form_url'),
-            'tour_colour_id' => $request->input('tour_colour_id'),
+            'tour_category_id' => $request->input('tour_category_id'),
             'is_active' => $request->input('is_active') === 'on' ? 1 : 0,
             'notes' => $request->input('notes'),
+            'invoice_footer' => $request->input('deposit'),
         ]);
         return redirect()->route('tours.view', ['tour' => $tour,]);
     }
 
     public function destroy(Tour $tour)
     {
+        if ($tour->orders()->count() > 0) {
+            return back()->withErrors(trans('custom.used-elsewhere', ['model' => 'Tour', 'parent' => 'ORder']));
+        }
         $tour->delete();
         return redirect()->route('tours.all');
     }
